+++ .gitignore
@@ -84,6 +84,7 @@ typings/
 # Nuxt.js build / generate output
 .nuxt
 dist
+*/dist
 
 # Gatsby files
 .cache/
