--- reactwp-be/components/rcf/model.js
@@ -1,14 +1,0 @@
-const { Model } = require('objection');
-const knex = require('../../knex/knex');
-
-Model.knex(knex);
-
-class Acf extends Model {
-
-    static get taleName() {
-        return 'acf';
-    }
-
-}
-
-module.exports = Acf;
