+++ reactwp-be/package.json
@@ -4,6 +4,8 @@
   "description": "",
   "main": "app.js",
   "scripts": {
+    "start": "tsc && node dist/src/app",
+    "dev": "nodemon -e js,ts --watch src --exec \"npm start\"",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
@@ -16,19 +18,14 @@
     "faker": "^4.1.0",
     "jsonwebtoken": "^8.5.1",
     "knex": "^0.20.8",
-    "knex-migrate-sql-file": "^1.0.3",
-    "mssql": "^6.0.1",
     "mysql": "^2.18.1",
-    "mysql2": "^2.1.0",
     "objection": "^2.1.2",
-    "oracle": "^0.4.1",
-    "pg": "^7.17.1",
     "slug-generator": "^1.0.6",
-    "slugify": "^1.3.6",
-    "sqlite3": "^4.1.1"
+    "slugify": "^1.3.6"
   },
   "devDependencies": {
-    "tslint": "^6.0.0",
+    "@types/express": "^4.17.3",
+    "tslint": "^6.1.0",
     "typescript": "^3.8.2"
   }
 }
