+++ reactwp-be/src/app.js
@@ -1,4 +1,5 @@
 const express = require('express');
+
 var bodyParser = require("body-parser");
 const cors = require('cors');
 const knex = require('./knex/knex')
