+++ reactwp-be/src/components/users/User.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const User = require('./model')
+const User = require('../../models/user')
 const bcrypt = require('bcrypt');
 const { generateToken } = require('./auth/auth')
