+++ reactwp-be/src/components/users/auth/auth.js
@@ -2,7 +2,7 @@ var jwt = require('jsonwebtoken');
 const express = require('express');
 const router = express.Router();
 const knex = require('../../../knex/knex')
-const User = require('../model')
+const User = require('../../../models/user')
 
 module.exports.generateToken = generateToken = (user) => {
     var u = {
