+++ reactwp-be/src/knex/knex.js
@@ -1,3 +1,3 @@
 const environment = process.env.ENVIRONMENT || 'development'
-const config = require('../knexfile.js')[environment];
+const config = require('../../knexfile.js')[environment];
 module.exports = require('knex')(config);
