+++ reactwp-be/src/models/pages.js
@@ -1,0 +1,24 @@
+const { Model } = require('objection');
+const knex = require('../knex/knex')
+
+Model.knex(knex)
+
+class Pages extends Model {
+    static get tableName() {
+        return 'post';
+    }
+
+    // static get relationMappings() {
+    //     return {
+    //         afc: {
+    //             relation: Model.HasManyRelation,
+    //             modelClass: Acf,
+    //             join: {
+    //                 from: ''
+    //             }
+    //         }
+    //     }
+    // }
+}
+
+module.exports = Pages;
