+++ reactwp-be/src/models/post.js
@@ -1,0 +1,20 @@
+const { Model } = require('objection');
+const knex = require('../knex/knex')
+
+Model.knex(knex);
+
+class Post extends Model {
+    static get tableName() {
+        return 'post'
+    }
+}
+
+
+// createSchema()
+// .then(() => knex.destroy())
+// .catch(err => {
+//     console.error(err);
+//     return knex.destroy();
+// });
+
+module.exports = Post;
