+++ reactwp-be/src/models/user.js
@@ -1,5 +1,5 @@
 const { Model } = require('objection');
-const knex = require('../../knex/knex')
+const knex = require('../knex/knex')
 
 Model.knex(knex);
