+++ reactwp-be/src/routes/admin.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 var jwt = require('jsonwebtoken');
 const { getTokenFromBearer } = require('../components/users/auth/auth')
-const pagesRoute = require('../components/pages/page');
+const pagesRoute = require('../components/post/page');
 const users = require('../components/users/user')
 const posts = require('../components/posts/posts')
 const postData = require('../components/postData/router')
