+++ reactwp-be/tsconfig.json
@@ -1,32 +1,24 @@
 {
   "compilerOptions": {
     "module": "commonjs",
-    "declaration": true,
     "removeComments": true,
     "allowJs": true,
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
     "target": "es6",
-    "sourceMap": true,
-    // "outDir": "./dist",
+    "outDir": "./dist/",
     "baseUrl": "./",
-    "incremental": true,
     "moduleResolution": "node",
     "lib": [
       "es2015"
     ],
-    "paths": {
-      "*": [
-        "node_modules/*",
-        "src/types/*"
-      ]
-    }
   },
   "exclude": [
-    "node_modules",
+    "./node_modules",
+    "./dist",
     "./knexfile.js"
   ],
   "include": [
-    "**/*"
+    "./src/**/*"
   ],
 }
